# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Daily Trends API",
        "description": "Esta es la documentación de la API.",
        "contact": {
            "email": "esparzaferran@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/feeds": {
            "get": {
                "tags": [
                    "Feeds"
                ],
                "summary": "Get all the articles",
                "operationId": "f852fc49018a9d36390d21e90f1d000d",
                "responses": {
                    "200": {
                        "description": "Lista de feeds obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Article"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "Content of the article"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-01-19T12:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-01-19T12:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Feeds"
                ],
                "summary": "Store article",
                "operationId": "a3c6b9596ee7fc33e562efb90148c1d6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "body of the article"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Nuevo artículo"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Contenido del artículo"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-19T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-19T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/feeds/{id}": {
            "get": {
                "tags": [
                    "Feeds"
                ],
                "summary": "Get an article by id",
                "operationId": "93d64c25c3d471aace28a9b999df098a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the feed",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feed encontrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Feed example"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Content of the feed"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-19T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-19T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Feeds"
                ],
                "summary": "Update article",
                "operationId": "88f5a871053539a63c4f687023302204",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the feed",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated article title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Contenido"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Article updated title"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Contenido updated"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-19T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-19T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Feeds"
                ],
                "summary": "Delete article",
                "operationId": "acd0827876b1e9d2827e83d5a1a6e04e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the feed",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article deleted"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Feeds",
            "description": "Feed model controller"
        }
    ]
}